Refactor: Extracting WithGuide
